[resample][documentation]
